requirements.txt
amundsen-databuilder==4.0.3
boto3==1.16.25
botocore==1.19.25
pyhocon==0.3.57
ruamel.yaml==0.16.12
sqlalchemy==1.3.22
psycopg2==2.8.6
